moved hooks folder into features, deleted redundent hook folder,
