+++ src/features/engine/twoOneGameEngine.ts
@@ -1,32 +1,10 @@
 import UciEngineWorker from "features/workers/stockfish";
 import { areMovesEqual, BoardOrientation, GameOverStates, getGameOverState, isPromoting, MoveAssignment, MoveWithAssignment } from "./chessEngine";
-import { Chess, ChessInstance, ShortMove } from 'chess.js';;
+import { Chess, ChessInstance, ShortMove } from 'chess.js';import { getRanElement } from "util/helpers";
+;
 
 const stockfishFile = "/default_stockfish/stockfish.js"
 const deadfishFile = "/deadfish/stockfish.js"
-/**
- * Gets a random element in array
- * @param {Array} array items An array containing the items.
- */
-function getRanElement(array: any[]) {
-    return array[Math.floor(Math.random() * array.length)]
-}
-
-/**
- * Shuffles array in place.
- * source: https://stackoverflow.com/questions/6274339/how-can-i-shuffle-an-array/6274381#6274381
- * @param {Array} a items An array containing the items.
- */
-function shuffle(a: any[]) {
-    var j, x, i;
-    for (i = a.length - 1; i > 0; i--) {
-        j = Math.floor(Math.random() * (i + 1));
-        x = a[i];
-        a[i] = a[j];
-        a[j] = x;
-    }
-    return a;
-}
 
 class TwoOneGameEngine {
     private _game: ChessInstance;
@@ -139,21 +117,12 @@ class TwoOneGameEngine {
             return new Promise(resolve => {
                 this.chessBotGood!.getMoves(this.game.history({verbose:true})).then(({calcMoves, bestMove}: any) => {
                     // Select bot moves
-                    const allMoves = this.game.moves({verbose: true});
                     const calcBestMove: MoveWithAssignment = { move: bestMove, assignment: MoveAssignment.best };
                     const randCalcMove: MoveWithAssignment = {
                         move: (calcMoves.length > 1 && calcMoves.find((m: { move: ShortMove; }) => !areMovesEqual(m.move, calcBestMove.move)).move)
                         || getRanElement(calcMoves).move, 
                         assignment: MoveAssignment.middle };
-    
-                    // randomly select a move
-                    const randomMove: MoveWithAssignment = {
-                        move: (allMoves.length > 1 && allMoves.find((m) => !areMovesEqual(m, calcBestMove.move) && !areMovesEqual(m, randCalcMove.move)))
-                        || getRanElement(allMoves),
-                        assignment: MoveAssignment.random };
-
-                    // Shuffle moves
-                    resolve(shuffle([randomMove, calcBestMove, randCalcMove]));
+                    resolve([calcBestMove, randCalcMove]);
               }).catch((msg) => {
                   console.log(msg);
               })
@@ -164,6 +133,7 @@ class TwoOneGameEngine {
         if (this.gameOn && this.isPlayerTurn()) {
             return new Promise(resolve => {
                 this.chessBotBad!.getMoves(this.game.history({verbose:true})).then(({calcMoves, bestMove}: any) => {
+                    console.log(calcMoves)
                     resolve(bestMove);
               }).catch((msg) => {
                   console.log(msg);
